- Corrección parcial de seguridad en logs con nivel WARNING y algunos con nivel INFO.

+++ afirma-signature-retriever/src/main/java/es/gob/afirma/signfolder/server/proxy/RetrieveService.java
@@ -22,6 +22,8 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import es.gob.afirma.signfolder.util.LoggerUtil;
+
 /** Servicio de almacenamiento temporal de firmas.
  * &Uacute;til para servir de intermediario en comunicaci&oacute;n entre JavaScript y aplicaciones nativas.
  * @author Tom&aacute;s Garc&iacute;a-Mer&aacute;s. */
@@ -106,26 +108,24 @@ public final class RetrieveService extends HttpServlet {
 		if (!inFile.isFile() || !inFile.canRead() || isExpired(inFile, RetrieveConfig.getExpirationTime())) {
 
 			if (!inFile.exists()) {
-				LOGGER.warning("El fichero con el identificador '" + id + "' no existe: " + inFile.getAbsolutePath()); //$NON-NLS-1$ //$NON-NLS-2$
+				LOGGER.warning("El fichero con el identificador '" + id + "' no existe: " + LoggerUtil.getCleanUserHomePath(inFile.getAbsolutePath())); //$NON-NLS-1$ //$NON-NLS-2$
 			}
 			else if (!inFile.isFile()) {
-				LOGGER.warning("El archivo con el identificador '" + id + "' no es un fichero: " + inFile.getAbsolutePath()); //$NON-NLS-1$ //$NON-NLS-2$
+				LOGGER.warning("El archivo con el identificador '" + id + "' no es un fichero: " + LoggerUtil.getCleanUserHomePath(inFile.getAbsolutePath())); //$NON-NLS-1$ //$NON-NLS-2$
 			}
 			else if (!inFile.canRead()) {
-				LOGGER.warning("El fichero con el identificador '" + id + "' no tiene permisos de lectura: " + inFile.getAbsolutePath()); //$NON-NLS-1$ //$NON-NLS-2$
+				LOGGER.warning("El fichero con el identificador '" + id + "' no tiene permisos de lectura: " + LoggerUtil.getCleanUserHomePath(inFile.getAbsolutePath())); //$NON-NLS-1$ //$NON-NLS-2$
 			}
 			else {
-				LOGGER.warning("El fichero con el identificador '" + id + "' esta caducado: " + inFile.getAbsolutePath()); //$NON-NLS-1$ //$NON-NLS-2$
+				LOGGER.warning("El fichero con el identificador '" + id + "' esta caducado: " + LoggerUtil.getCleanUserHomePath(inFile.getAbsolutePath())); //$NON-NLS-1$ //$NON-NLS-2$
 			}
 
 			out.println(
 				ErrorManager.genError(ErrorManager.ERROR_INVALID_DATA_ID)  + " ('" + id + "')" //$NON-NLS-1$ //$NON-NLS-2$
 			);
 			// Que el fichero sea de tipo fichero, implica que existe
-			if (inFile.isFile()) {
-				if (!RetrieveConfig.DEBUG) {
-					inFile.delete();
-				}
+			if (inFile.isFile() && !RetrieveConfig.DEBUG) {
+				inFile.delete();
 			}
 		}
 		else {
@@ -167,7 +167,7 @@ public final class RetrieveService extends HttpServlet {
 				catch(final Exception e) {
 					// Suponemos que el fichero ha sido eliminado por otro hilo
 					LOGGER.warning(
-						"No se ha podido eliminar el fichero '" + file.getAbsolutePath() + "', es probable que se elimine en otro hilo de ejecucion: " + e //$NON-NLS-1$ //$NON-NLS-2$
+						"No se ha podido eliminar el fichero '" + LoggerUtil.getCleanUserHomePath(file.getAbsolutePath()) + "', es probable que se elimine en otro hilo de ejecucion: " + e //$NON-NLS-1$ //$NON-NLS-2$
 					);
 				}
 			}
